07/16現在

+++ JC21PS/src/jp/co/jcps/A06/JoinApprovalSaveServlet.java
@@ -43,7 +43,7 @@ public class JoinApprovalSaveServlet extends HttpServlet {
 		// リクエストから情報を取得する
 		boolean approvalFlg = request.getParameter("approvalFlg").equals("true");
 		// TODO: リクエストから承認・否認するユーザーのユーザーIDを取得しなさい
-		String registUserId = ;
+		String registUserId = request.getParameter("user_ID");
 
 		// セッションからログイン中のユーザーの部長クラブIDを取得する
 		String leaderClubId = (String) request.getSession().getAttribute("leaderClubId");
@@ -52,11 +52,11 @@ public class JoinApprovalSaveServlet extends HttpServlet {
 			if (approvalFlg) {
 				// 承認する場合
 				// TODO: 部員登録申請を承認する場合のみ実行する処理（メソッド）を呼び出しなさい。
-
+				saveClubMember(leaderClubId,registUserId);
 
 			}
 			//TODO: 部員登録申請を承認する場合および拒否する場合どちらも実行する処理（メソッド）を呼び出しなさい。
-
+				deleteJoinRequest(leaderClubId,registUserId);
 		}catch(Exception e) {
 			request.getRequestDispatcher("ERROR/Error.jsp").forward(request, response);
 		}
@@ -74,7 +74,7 @@ public class JoinApprovalSaveServlet extends HttpServlet {
 
 		//SQLを宣言
 		// TODO: SQL文を完成させなさい。
-		String sql = "INSERT INTO ";
+		String sql = "INSERT INTO TRN_CLUB_MEMBER('CLUB_ID','USER_ID') VALUES('registClubId','registUserId')";
 
 		// SQLに埋め込むパラメータリストを定義
 		List<String> paramList = new ArrayList<String>();
@@ -97,7 +97,7 @@ public class JoinApprovalSaveServlet extends HttpServlet {
 
 		//SQLを宣言
 		// TODO: SQL文を完成させなさい。
-		String sql = "DELETE FROM ";
+		String sql = "DELETE FROM TRN_JOIN_REQUEST WHERE CLUB_ID=registClubId AND USER_ID=registUserId";
 
 		// SQLに埋め込むパラメータリストを定義
 		List<String> paramList = new ArrayList<String>();
